Deleted Doors, Set up Main Game Events, Lots of Code!
Basis for game events are set up- player can pick up notes which trigger events. game checks if player has picked up the total amount of notes per room. the code also checks for "lamp conditions", and shows a UI overlay of the note. When the overlay is showing, the player cannot move. there's an establishing cutscene, and the doors for all of the enterable rooms are now deleted.